✨ feat(taskx): boilerplate for dataview and tasks
Prepare integration with plugins Dataview and Tasks.

We need to preprocess "@codemirror/language" to inhibit the prepare stage or
install can't complete (dep is introduced by "obsidian-dataview")

Currently the script is just an hello world example ("Do this", "Do that!")
to verify our plugin can write to a dataviewjs block.

We rename our plugin "obsidian-taskx", our 'install' script 'copy' and we update
the README.